mission clear

+++ 07_push_swap/libft/ft_striteri.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_striteri.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: yoson <yoson@student.42.fr>                +#+  +:+       +#+        */
+/*   By: suhwpark <suhwpark@student.42seoul.>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/07/09 22:02:26 by yoson             #+#    #+#             */
-/*   Updated: 2022/10/11 19:33:35 by yoson            ###   ########.fr       */
+/*   Created: 2022/11/15 15:36:04 by suhwpark          #+#    #+#             */
+/*   Updated: 2022/11/23 19:16:07 by suhwpark         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_striteri(char *s, void (*f)(unsigned int, char*))
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
